requirements/base.txt
Django==4.2.2
django-compressor==4.3.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.2
django-jalali-date==1.0.2
django-jazzmin==2.6.0
django-lifecycle==1.0.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
drf-yasg==1.21.6
psycopg2==2.9.6
elastic-apm==6.16.2
Faker==18.10.1
celery==5.3.0

